% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{AnglePlot}
\alias{AnglePlot}
\title{Plot Angle}
\usage{
AnglePlot(
  DataTable,
  Angle,
  CoordRef = NULL,
  ObjectTable = NULL,
  colourScheme = "dark",
  ObjectHighlight = "alpha",
  FacetRef = NULL
)
}
\arguments{
\item{DataTable}{A data.table used as input.}

\item{Angle}{A string referencing the angle.}

\item{CoordRef}{A string referencing coordinate columns.}

\item{ObjectTable}{Optional data.table with Objects.}

\item{colourScheme}{A string indicating the colour scheme (default = "dark").}

\item{ObjectHighlight}{A string indicating if the reference object should be highlighted (default = "alpha", other options: "colour", "stroke", "none" or colour valid string).}

\item{FacetRef}{A string as reference for facet wrap (for multi-view).}
}
\value{
Returns a ggplot.
}
\description{
This function plots the speed of the object used as speed reference. When adding the x/y 
references a 2D-plot is drawn using colour-coding for the speed. If the x/y information 
is omitted the speed will be plotted against time.
}
