\name{BehaviouR-package}
\alias{BehaviouR-package}
\alias{BehaviouR}
\docType{package}
\title{\packageTitle{BehaviouR}}
\description{\packageDescription{BehaviouR}}
\section{Package Content}{\packageIndices{BehaviouR}}
\author{\packageAuthor{BehaviouR}}
\section{Maintainer}{\packageMaintainer{BehaviouR}}
\keyword{package}
