% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{CustomColourPalette}
\alias{CustomColourPalette}
\title{Colour scheme}
\usage{
CustomColourPalette(Mode = "dark", n = 5)
}
\arguments{
\item{Mode}{A string for the colour scheme ("dark", "light").}

\item{n}{An interger indicating length of output.}
}
\value{
Returns a vector string containing colours in hex-code.
}
\description{
This function generates colourblind friendly schemes and returns them as hex-string.
}
