% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedDistCumDist.R
\name{DistSpeedCalc}
\alias{DistSpeedCalc}
\title{Function to calculate Speed/Distance/Cumulative distance}
\usage{
DistSpeedCalc(CoordTable, SpeedRef, Interval)
}
\arguments{
\item{CoordTable}{A table including coordinates of labels.}

\item{SpeedRef}{A string indicating the label used for computing speed.}

\item{Interval}{A double corresponding to sampling rate.}
}
\value{
Returns a data.frame which includes the labels and the directory to the folders.
}
\description{
This function will get the directory for any given label by opening a user interaction.
}
