% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{LengthPlot}
\alias{LengthPlot}
\title{Plot Length}
\usage{
LengthPlot(
  DataTable,
  Length,
  x = NULL,
  y = NULL,
  ObjectTable = NULL,
  Unit = "px",
  ColourFlip = T
)
}
\arguments{
\item{DataTable}{A data.table used as input.}

\item{Length}{A string referencing the Length column.}

\item{x}{A string referencing the x-cordinate column.}

\item{y}{A string referencing the y-cordinate column.}

\item{ObjectTable}{Optional data.table with Objects.}

\item{Unit}{A string indicating the speed unit (default = "px").}

\item{ColourFlip}{A bool indicating the whether colour scheme should be inverted (default = TRUE).}
}
\value{
Returns a ggplot.
}
\description{
This function plots the Length of the object used as Length reference. When adding the x/y 
references a 2D-plot is drawn using colour-coding for the Length If the x/y inforamtion 
is omitted the Length will be plotted against time. The length of a vector can be used to 
estimate rearing or other postures.
}
