% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{LocationPlot}
\alias{LocationPlot}
\title{Plot Location}
\usage{
LocationPlot(
  CoordTable,
  CoordRef = NULL,
  ObjectTable = NULL,
  Density = TRUE,
  BinNumber = 500,
  FacetRef = NULL
)
}
\arguments{
\item{CoordTable}{A data.table used as input.}

\item{CoordRef}{A string referencing coordinate columns.}

\item{ObjectTable}{Optional data.table with Objects.}

\item{Density}{A bool indicating whether the density should be plotted.}

\item{BinNumber}{An integer indicating the number of bins for the density (default = 500).}

\item{FacetRef}{A string as reference for facet wrap (for multi-view).}
}
\value{
Returns a ggplot.
}
\description{
This function plots the speed of the object used as speed reference. When adding the x/y 
references a 2D-plot is drawn using colour-coding for the speed. If the x/y information 
is omitted the speed will be plotted against time.
}
