% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFunctions.R
\name{ObjectAngle}
\alias{ObjectAngle}
\title{Calculate angle to objects}
\usage{
ObjectAngle(
  CoordTable,
  ObjectTable,
  Ref,
  RefStart = NULL,
  ReferenceColumn = "",
  Overwrite = TRUE
)
}
\arguments{
\item{CoordTable}{A table including coordinates of labels.}

\item{ObjectTable}{A table including the objects.}

\item{Ref}{A string indicating the label used for object angle.}

\item{RefStart}{A string indicating the label used for object angle.}

\item{ReferenceColumn}{A string indicating the reference column (used for batch analysis).}

\item{Overwrite}{A bool indicating if output should be overwritten if it exists already (default = TRUE).}
}
\value{
Modifies existing DataTable to add the angle to an object.
}
\description{
This function calculates the angle from a reference to objects from an object table and adds the resulting distances to the DataTable as columns.
}
