% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFunctions.R
\name{ObjectDistance}
\alias{ObjectDistance}
\title{Calculate distance to objects}
\usage{
ObjectDistance(
  CoordTable,
  ObjectTable,
  Ref,
  ReferenceColumn = "",
  Overwrite = TRUE
)
}
\arguments{
\item{CoordTable}{A table including coordinates of labels.}

\item{ObjectTable}{A table including the objects.}

\item{Ref}{A string indicating the label used for object distance.}

\item{ReferenceColumn}{A string indicating the reference column (used for batch analysis).}

\item{Overwrite}{A bool indicating if output should be overwritten if it exists already (default = TRUE).}
}
\value{
Modifies existing DataTable.
}
\description{
This function calculates the distance from a reference to objects from an object table and adds the resulting distances to the DataTable as columns.
}
