% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{SpeedPlot}
\alias{SpeedPlot}
\title{Plot Speed}
\usage{
SpeedPlot(
  DataTable,
  Speed,
  x = NULL,
  y = NULL,
  ObjectTable = NULL,
  Unit = "px/s"
)
}
\arguments{
\item{DataTable}{A data.table used as input.}

\item{Speed}{A string referencing the speed column.}

\item{x}{A string referencing the x-coordinate column.}

\item{y}{A string referencing the y-coordinate column.}

\item{ObjectTable}{Optional data.table with Objects.}

\item{Unit}{A string indicating the speed unit (default = "px/s").}
}
\value{
Returns a ggplot.
}
\description{
This function plots the speed of the object used as speed reference. When adding the x/y 
references a 2D-plot is drawn using colour-coding for the speed. If the x/y inforamtion 
is omitted the speed will be plotted against time.
}
