% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{StartApp}
\alias{StartApp}
\title{Start Analysis Application}
\usage{
StartApp(Cores = NULL)
}
\arguments{
\item{Cores}{Optional input argument to set number of threads. Default is set by data.table.}
}
\description{
This function calls the Behaviour user interface which allows loading DeepLabCut output, analysing, plotting, and exporting data.
}
