% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maintenance.R
\name{VariableNameCheck}
\alias{VariableNameCheck}
\title{Name checking}
\usage{
VariableNameCheck(DataTable, NameString)
}
\arguments{
\item{DataTable}{A table including coordinates of labels.}

\item{NameString}{A string indicating the label used for object distance.}
}
\value{
Modifies existing DataTable.
}
\description{
This function checks for duplicates of columns.
}
