% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFunctions.R
\name{ZoneEntry}
\alias{ZoneEntry}
\title{Calculate entry to object}
\usage{
ZoneEntry(
  CoordTable,
  DistanceRef,
  Length,
  AngleInclusion = TRUE,
  AngleRef = NULL,
  AngleRange = pi * (40/360),
  Overwrite = TRUE
)
}
\arguments{
\item{CoordTable}{A table including coordinates of labels.}

\item{DistanceRef}{A string indicating the distance reference column.}

\item{Length}{double as distance cut-off for zone.}

\item{AngleInclusion}{A bool indicating if angles should be considered for entry detection.}

\item{AngleRef}{A string indicating angle reference column.}

\item{AngleRange}{A double as indicating the angle range (+/-).}

\item{Overwrite}{A bool indicating if output should be overwritten if it exists already (default = TRUE).}
}
\value{
Modifies existing DataTable and add the binary output for an entry, .
}
\description{
This function calculates the angle from a reference to objects from an object table and adds the resulting distances to the DataTable as columns.
}
