% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centroid.R
\name{AddCentroid}
\alias{AddCentroid}
\title{Calculate centroid}
\usage{
AddCentroid(
  CoordTable,
  CornerNames,
  ReferenceColumn,
  OutputName,
  Overwrite = TRUE
)
}
\arguments{
\item{CoordTable}{DataTable with coordinates.}

\item{CornerNames}{A vector string with labels .}

\item{ReferenceColumn}{A string indicating the reference column.}

\item{OutputName}{A string for output.}

\item{Overwrite}{A bool indicating if ouput should be overwritten if it exists already (default = TRUE).}
}
\value{
Add centroid coordinates
}
\description{
This function calculates the centroid of object with .
}
