% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Angles.R
\name{AngleCalc}
\alias{AngleCalc}
\title{Calculate angle between two vectors}
\usage{
AngleCalc(CoordTable, VectorStart, VectorEnd, OutputName)
}
\arguments{
\item{CoordTable}{A table including coordinates of labels.}

\item{VectorStart}{A table including the objects.}

\item{VectorEnd}{A vector string indicating the labels used for computing object angle.}

\item{OutputName}{A string indicating the label used for object angle.}
}
\value{
Modifies existing DataTable.
}
\description{
This function calculates the angle from a reference to objects from an object table and adds the resulting distances to the DataTable as columns.
}
