% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{AnglePlot}
\alias{AnglePlot}
\title{Plot Angle}
\usage{
AnglePlot(
  DataTable,
  Angle,
  x = NULL,
  y = NULL,
  ObjectTable = NULL,
  colourScheme = "dark",
  ObjectHighlight = "alpha"
)
}
\arguments{
\item{DataTable}{A data.table used as input.}

\item{x}{A string referencing the x-ccordinate column.}

\item{y}{A string referencing the y-ccordinate column.}

\item{ObjectTable}{Optional data.table with Objects.}

\item{colourScheme}{A string indicating the colour scheme (default = "dark").}

\item{ObjectHighlight}{A string indicating if the reference object should be highlighted (default = "alpha", other options: "colour", "stroke", "none" or colour valid string).}
}
\value{
Returns a ggplot.
}
\description{
This function plots the speed of the object used as speed reference. When adding the x/y 
references a 2D-plot is drawn using colour-coding for the speed. If the x/y inforamtion 
is omitted the speed will be plotted against time.
}
