% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centroid.R
\name{CentroidCollect}
\alias{CentroidCollect}
\title{Wrapper for centroid calculation}
\usage{
CentroidCollect(CoordTable, MouseLabels, Overwrite = TRUE)
}
\arguments{
\item{CoordTable}{DataTable with coordinates.}

\item{MouseLabels}{A vector string with labels.}

\item{Overwrite}{A bool indicating if ouput should be overwritten if it exists already (default = TRUE).}
}
\value{
Add centroid coordinates
}
\description{
This function calculates the centroid of object with .
}
