% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoad.R
\name{DeepLabCutLoad}
\alias{DeepLabCutLoad}
\title{This function generates a DataTable from a csv exported with DeepLabCut and computes parameters as Speed/Distance, and other object related features.}
\usage{
DeepLabCutLoad(
  FileName,
  FrameRate,
  MouseLabels,
  ObjectLabels,
  ObjectNumber,
  xScale = 1,
  yScale = 1,
  JumpCorrections = T,
  interpWindow = 21,
  includeAll = T
)
}
\arguments{
\item{FileName}{Name of CSV file including the path.}

\item{FrameRate}{A double indicating the frame rate.}

\item{MouseLabels}{A vector string indicating the labels used.}

\item{ObjectLabels}{A string indicating the label used for object.}

\item{ObjectNumber}{An integer indicating the number of objects.}

\item{xScale}{A double representing the scaling factor.}

\item{yScale}{A double representing the scaling factor.}

\item{JumpCorrections}{A bool indicating correction of label jump artefacts.}

\item{interpWindow}{An integer determining the length of spline fitting area around artefact.}

\item{includeAll}{A bool indicating if other labels should be imported as well (default = TRUE).}
}
\value{
Generate DataTable from CSV file
}
\description{
This function generates a DataTable from a csv exported with DeepLabCut and computes parameters as Speed/Distance, and other object related features.
}
