% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{DistancePlot}
\alias{DistancePlot}
\title{Plot Distance}
\usage{
DistancePlot(
  CoordTable,
  Distance,
  CoordRef = NULL,
  ObjectTable = NULL,
  Unit = "px",
  ObjectDistance = F,
  FacetRef = NULL
)
}
\arguments{
\item{CoordTable}{A data.table used as input.}

\item{Distance}{A string referencing the distance column.}

\item{CoordRef}{A string referencing coordinate columns.}

\item{ObjectTable}{Optional data.table with Objects.}

\item{Unit}{A string indicating the speed unit (default = "px/s").}

\item{ObjectDistance}{A bool indicating if reference is linked to stationary object(s) (default = FALSE).}

\item{FacetRef}{A string as reference for facet wrap (for multi-view).}
}
\value{
Returns a ggplot.
}
\description{
This function plots the distance of the object used as distance reference. When adding the x/y 
references a 2D-plot is drawn using colour-coding for the distance. If the x/y information 
is omitted the distance will be plotted against time.
}
