% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedDistCumDist.R
\name{DistSpeedCalc}
\alias{DistSpeedCalc}
\title{Speed/Distance/Cumulative distance}
\usage{
DistSpeedCalc(CoordTable, SpeedRef, Interval)
}
\arguments{
\item{CoordTable}{A table including coordinates of labels.}

\item{SpeedRef}{A string indicating the label used for computing speed.}

\item{Interval}{A double corresponding to sampling rate.}
}
\value{
Modifies existing DataTable.
}
\description{
This function calculates the speed, distance, and cumulative distance and add them as DataTable column.
}
