% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{LengthPlot}
\alias{LengthPlot}
\title{Plot Length}
\usage{
LengthPlot(
  CoordTable,
  Length,
  CoordRef = NULL,
  ObjectTable = NULL,
  Unit = "px",
  ColourFlip = T,
  FacetRef = NULL
)
}
\arguments{
\item{CoordTable}{A data.table used as input.}

\item{Length}{A string referencing the Length column.}

\item{CoordRef}{A string referencing coordinate columns.}

\item{ObjectTable}{Optional data.table with Objects.}

\item{Unit}{A string indicating the speed unit (default = "px").}

\item{ColourFlip}{A bool indicating the whether colour scheme should be inverted (default = TRUE).}

\item{FacetRef}{A string as reference for facet wrap (for multi-view).}
}
\value{
Returns a ggplot.
}
\description{
This function plots the Length of the object used as Length reference. When adding the x/y 
references a 2D-plot is drawn using colour-coding for the Length If the x/y information 
is omitted the Length will be plotted against time. The length of a vector can be used to 
estimate rearing or other postures.
}
