% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{LocationPlot}
\alias{LocationPlot}
\title{Plot Location}
\usage{
LocationPlot(
  DataTable,
  x = NULL,
  y = NULL,
  ObjectTable = NULL,
  Density = TRUE,
  BinNumber = 500
)
}
\arguments{
\item{DataTable}{A data.table used as input.}

\item{x}{A string referencing the x-coordinate column.}

\item{y}{A string referencing the y-coordinate column.}

\item{ObjectTable}{Optional data.table with Objects.}

\item{Density}{A bool indicating whether the density should be plotted.}

\item{BinNumber}{An integer indicating the number of bins for the density (default = 500).}
}
\value{
Returns a ggplot2::ggplot.
}
\description{
This function plots the speed of the object used as speed reference. When adding the x/y 
references a 2D-plot is drawn using colour-coding for the speed. If the x/y information 
is omitted the speed will be plotted against time.
}
