% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFunctions.R
\name{ObjectAngle}
\alias{ObjectAngle}
\title{Calculate angle to objects}
\usage{
ObjectAngle(CoordTable, ObjectTable, ObjectLabels, Ref)
}
\arguments{
\item{CoordTable}{A table including coordinates of labels.}

\item{ObjectTable}{A table including the objects.}

\item{ObjectLabels}{A vector string indicating the labels used for computing object angle.}

\item{Ref}{A string indicating the label used for object angle.}
}
\value{
Modifies existing DataTable.
}
\description{
This function calculates the angle from a reference to objects from an object table and adds the resulting distances to the DataTable as columns.
}
