% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFunctions.R
\name{ObjectDistance}
\alias{ObjectDistance}
\title{Calculate distance to objects}
\usage{
ObjectDistance(CoordTable, ObjectTable, ObjectLabels, Ref)
}
\arguments{
\item{CoordTable}{A table including coordinates of labels.}

\item{ObjectTable}{A table including the objects.}

\item{ObjectLabels}{A vector string indicating the labels used for computing object distances.}

\item{Ref}{A string indicating the label used for object distance.}
}
\value{
Modifies existing DataTable.
}
\description{
This function calculates the distance from a reference to objects from an object table and adds the resulting distances to the DataTable as columns.
}
