% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{SpeedPlot}
\alias{SpeedPlot}
\title{Plot Speed}
\usage{
SpeedPlot(
  DataTable,
  Speed,
  CoordRef = NULL,
  ObjectTable = NULL,
  Unit = "px/s",
  FacetRef = NULL
)
}
\arguments{
\item{DataTable}{A data.table used as input.}

\item{Speed}{A string referencing the speed column.}

\item{CoordRef}{A string referencing coordinate columns.}

\item{ObjectTable}{Optional data.table with Objects.}

\item{Unit}{A string indicating the speed unit (default = "px/s").}

\item{FacetRef}{A string as reference for facet wrap (for multi-view).}
}
\value{
Returns a ggplot.
}
\description{
This function plots the speed of the object used as speed reference. When adding the x/y 
references a 2D-plot is drawn using colour-coding for the speed. If the x/y information 
is omitted the speed will be plotted against time.
}
