% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{StartApp}
\alias{StartApp}
\title{Start Analysis Application}
\usage{
StartApp()
}
\description{
This function calls the Behaviour user interface which allows loading DeepLabCut output, analysing and plotting data.
}
