% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorLength.R
\name{VectorLength}
\alias{VectorLength}
\title{Vector Length}
\usage{
VectorLength(CoordTable, VectorStart, VectorEnd, OutputName, Overwrite = TRUE)
}
\arguments{
\item{CoordTable}{A table including coordinates of labels.}

\item{VectorStart}{A string indicating the label used for start of vector.}

\item{VectorEnd}{A string indicating the label used for end of vector.}

\item{OutputName}{A string indicating the label used as output.}

\item{Overwrite}{A bool indicating if output should be overwritten if it exists already (default = TRUE).}
}
\value{
Modifies existing DataTable.
}
\description{
This function calculates the vector length of two points at any given point in time.
}
